Add business-specific notification settings
